+++ tests/test_equil.cpp
@@ -38,7 +38,7 @@
 #include <opm/core/pressure/msmfem/partition.h>
 
 #include <opm/core/utility/parameters/ParameterGroup.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 
 #include <array>
 #include <iostream>
